fixed - now just need to implement SequenceMatcher

+++ src/string_matcher.rs
@@ -7,8 +7,8 @@ impl<'a> SequenceMatcher<'a> {
     pub fn new(s1: &'a str, s2: &'a str) -> SequenceMatcher<'a> {
         SequenceMatcher{ shorter:s1, longer:s2 }
     }
-    pub fn get_matching_blocks(&self) -> Vec<(u32,u32,u32)> {
-        let result: Vec<(u32, u32, u32)> = Vec::new();
+    pub fn get_matching_blocks(&self) -> Vec<(usize,usize,usize)> {
+        let result: Vec<(usize,usize,usize)> = Vec::new();
 
         result
     }
